fix a hacky check so it will work (still hacky)
The check attempts to detect when m_sortingcol has been (wrongly) decremented below zero;
without this fix the check is ineffective (can never fail) because m_sortingcol is uint32_t,
so doesn't sign-extend when cast to int64_t, so the cast result could never be negative.
Changing the code to cast to (signed) without changing the length avoids this problem.
-       if((int64_t)m_sortingcol < 0)
+       if((int32_t)m_sortingcol < 0)

+++ userspace/libsinsp/viewinfo.cpp
@@ -183,7 +183,7 @@ void sinsp_view_info::set_sorting_col()
 		throw sinsp_exception("view format error: more than one sorting column");
 	}
 
-	if((int64_t)m_sortingcol < 0)
+	if((int32_t)m_sortingcol < 0)
 	{
 		ASSERT(false);
 		throw sinsp_exception("view sorting column configuration error");
